Fix partition_values, highest_prime_sieve, and fix overflow in mod_1 algs

+++ delta/modulo_2.h
@@ -46,7 +46,7 @@ uint32_t delta_modulo_2(Factors f) {
         }
 
         // The slow intuitive way:
-            for (int p = p_start; p < p_end; p += step * r) {
+            for (int p = p_start; p < (int64_t)p_end; p += step * r) {
                 cout << p << endl;
                 int i = (p - evens_max) / 2;     // index after translation and scaling
                 evens_visited[i] = true;
